# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.LabRequest:
    properties:
      address:
        type: string
      email:
        type: string
      lab_name:
        type: string
      phone:
        type: string
      registration_no:
        type: string
      report_footer:
        type: string
      tagline:
        type: string
    required:
    - lab_name
    - registration_no
    type: object
  github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.LabResponse:
    properties:
      address:
        type: string
      email:
        type: string
      id:
        type: string
      lab_name:
        type: string
      phone:
        type: string
      registration_no:
        type: string
      report_footer:
        type: string
      tagline:
        type: string
      updated_by:
        type: string
    type: object
...
  github_com_jenish-brainztechs_go-backend_internal_core_domain.LabSettings:
    properties:
      address:
        type: string
      email:
        type: string
      id:
        type: string
      labName:
        type: string
      phone:
        type: string
      registrationNo:
        type: string
      reportFooter:
        type: string
      tagline:
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
...
  internal_adapter_handler_http.errorResponse:
    properties:
      messages:
        example:
        - Error message 1
        - ' Error message 2'
        items:
          type: string
        type: array
      success:
        example: false
        type: boolean
    type: object
  internal_adapter_handler_http.response:
    properties:
      data: {}
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Laboratory Information System API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: LIS Backend API
  version: "1.0"
paths:
...
  /admin/lab:
    get:
      consumes:
      - application/json
      description: Get all lab settings (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_adapter_handler_http.response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.LabResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: List all lab settings
      tags:
      - Lab Settings
    post:
      consumes:
      - application/json
      description: Create new lab settings (admin only)
      parameters:
      - description: Lab settings details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.LabRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_adapter_handler_http.response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_core_domain.LabSettings'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: Create lab settings
      tags:
      - Lab Settings
...
  /admin/lab/{id}:
    patch:
      consumes:
      - application/json
      description: Update lab settings by ID (admin only)
      parameters:
      - description: Lab ID
        in: path
        name: id
        required: true
        type: string
      - description: Lab settings details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_adapter_handler_http_dto.LabRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: Update lab settings
      tags:
      - Lab Settings
...
  /lab/get/{id}:
    get:
      consumes:
      - application/json
      description: Get lab settings by ID
      parameters:
      - description: Lab ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_adapter_handler_http.response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_jenish-brainztechs_go-backend_internal_core_domain.LabSettings'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler_http.errorResponse'
      security:
      - BearerAuth: []
      summary: Get lab settings
      tags:
      - Lab Settings
...
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"